Added additional robodata; added E column
Added robodata2.csv, which is similar to robodata.csv but more extensive. Contains 292 poses, more extensive joint positions (particularly in A1), and motion in the E axis, which is motion of the entire robot on a linear track aligned with A1=0. Modified structure of robodata to also have an E column with all values set to 0, as that robot doesn't have a track.