Split accuracy analysis and model derivation into separate notebooks
It got way too long. It's still way too long. But at least now it's better sub-divided.
I fully went through the entire accuracy analysis portion and updated it to use 7 axes instead of 6, as well as doing a lot of refactoring, proofreading, fine-tuning, explaining, and so-on. The accuracy analysis is now effectively complete, unless I come up with more things I want to do to it.